backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic-settings==2.5.2
python-multipart==0.0.9
structlog==24.1.0
pytest==8.0.2
httpx==0.27.0
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pypdf==4.1.0
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pytesseract==0.3.10
Pillow>=10.2.0
numpy<2
sentence-transformers==2.6.1
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0
pinecone==5.0.1
google-genai>=1.0.0
supermemory
